pyproject.toml
[project]
name = "8edr1-74451"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pre-commit>=4.2.0",
    "psycopg2>=2.9.10",
    "ruff>=0.12.8",
    "model-bakery>=1.20.5",
]
